moved senior interface to a folder
